% Generated by build script
\name{cagov_dataset}
\alias{cagov_dataset}
\title{Get detailed info about a single California dataset}
\usage{
cagov_dataset()
}
\arguments{
\item{dataset_name}{Dataset name or ID}
}
\value{tibble: id, name, title, organization, author, maintainer,}
\description{Get detailed info about a single California dataset}
