% Generated by build script
\name{cagov_datastore}
\alias{cagov_datastore}
\title{Query California datastore}
\usage{
cagov_datastore()
}
\arguments{
\item{resource_id}{Resource ID (get from cagov_resources())}
\item{filters}{Named list of field=value filters}
\item{query}{Full-text search within the resource}
\item{limit}{Max rows (default 100)}
\item{offset}{Pagination offset (default 0)}
}
\value{tibble of query results}
\description{Query California datastore}
