% Generated by build script
\name{cagov_organizations}
\alias{cagov_organizations}
\title{List organizations on California Open Data}
\usage{
cagov_organizations()
}
\arguments{
\item{limit}{Max organizations (default 50)}
}
\value{tibble: id, name, title, package_count}
\description{List organizations on California Open Data}
