% Generated by build script
\name{cagov_resources}
\alias{cagov_resources}
\title{Get resources for a California dataset}
\usage{
cagov_resources()
}
\arguments{
\item{dataset_name}{Dataset name or ID}
}
\value{tibble: id, name, format, url, description}
\description{Get resources for a California dataset}
