% Generated by build script
\name{cagov_search}
\alias{cagov_search}
\title{Search California Open Data datasets}
\usage{
cagov_search()
}
\arguments{
\item{query}{Search term}
\item{rows}{Number of results (default 50)}
\item{start}{Offset (default 0)}
}
\value{tibble: id, name, title, organization, num_resources, notes}
\description{Search California Open Data datasets}
